+++ morecantile/scripts/cli.py
@@ -11,6 +11,8 @@ import morecantile
 
 logger = logging.getLogger(__name__)
 
+WGS84_CRS = CRS.from_epsg(4326)
+
 
 def configure_logging(verbosity):
     """Configure log verbosity.
@@ -189,6 +191,11 @@ def cli(ctx, verbose, quiet):
     help="Path to TileMatrixSet JSON file.",
     type=click.Path(),
 )
+@click.option(
+    "--crs",
+    help="Geographic CRS. Default to WGS84.",
+    type=str,
+)
 @click.pass_context
 def shapes(  # noqa: C901
     ctx,
@@ -204,6 +211,7 @@ def shapes(  # noqa: C901
     extents,
     buffer,
     tms,
+    crs,
 ):
     """
     Reads one or more Web Mercator tile descriptions
@@ -259,6 +267,7 @@ def shapes(  # noqa: C901
             projected=projected,
             buffer=buffer,
             precision=precision,
+            geographic_crs=CRS.from_user_input(crs) if crs else WGS84_CRS,
         )
         bbox = feature["bbox"]
         w, s, e, n = bbox
@@ -445,8 +454,6 @@ def custom(
     epsg, extent, name, minzoom, maxzoom, tile_width, tile_height, extent_epsg, title
 ):
     """Create Custom TMS."""
-    extent_crs = CRS.from_epsg(extent_epsg) if extent_epsg else None
-
     tms = morecantile.TileMatrixSet.custom(
         extent,
         CRS.from_epsg(epsg),
@@ -455,7 +462,7 @@ def custom(
         maxzoom=maxzoom,
         tile_width=tile_width,
         tile_height=tile_height,
-        extent_crs=extent_crs,
+        extent_crs=CRS.from_epsg(extent_epsg) if extent_epsg else None,
         title=title or "Custom TileMatrixSet",
     )
     click.echo(tms.json(exclude_none=True))
@@ -526,6 +533,11 @@ def custom(
     default=None,
     help="Shift shape x and y values by a constant number",
 )
+@click.option(
+    "--crs",
+    help="Geographic CRS. Default to WGS84.",
+    type=str,
+)
 def tms_to_geojson(  # noqa: C901
     input,
     level,
@@ -538,6 +550,7 @@ def tms_to_geojson(  # noqa: C901
     collect,
     extents,
     buffer,
+    crs,
 ):
     """Print TMS document as GeoJSON."""
     tms = morecantile.TileMatrixSet(**json.load(input))
@@ -568,6 +581,7 @@ def tms_to_geojson(  # noqa: C901
                 projected=projected,
                 buffer=buffer,
                 precision=precision,
+                geographic_crs=CRS.from_user_input(crs) if crs else WGS84_CRS,
             )
             bbox = feature["bbox"]
             w, s, e, n = bbox