+++ CHANGES.md
@@ -1,7 +1,52 @@
 
 ## Unreleased
 
+* use `WGS84` as default CRS for `TileMatrixSet.feature` GeoJSON response (as per specification)
+* add `geographic_crs` option for `TileMatrixSet.feature` method
+* update non-WGS84 CRS notation in `TileMatrixSet.feature` GeoJSON response
+
+    ```python
+    # before
+    "properties": {
+        "crs": {
+            "type": "EPSG",
+            "properties": {"code": 3857},
+        }
+    }
+
+    # now
+    "properties": {
+        "crs": {
+            "type": "name",
+            "properties": {"name": "http://www.opengis.net/def/crs/EPSG/0/3857"},
+        }
+        # or
+        "crs": {
+            "type": "wkt",
+             "properties": {"wkt": "..."}},
+        }
+    }
+    ```
+
+* rename `grid_name -> tms` and `grid_crs -> tms_crs` property names in `TileMatrixSet.feature` GeoJSON response
 * remove python 3.8 support
+* check tile's zoom against TMS's `maxzoom` in `TileMatrixSet.is_valid` and add `strict=True|False` options
+
+    ```python
+    import morecantile
+    tms = morecantile.tms.get("WebMercatorQuad")
+
+    # before
+    assert tms.is_valid(0, 0, 25)
+    >> UserWarning: TileMatrix not found for level: 25 - Creating values from TMS Scale.
+
+    # now
+    assert tms.is_valid(0, 0, 25), "Tile(0, 0, 25) is not valid"
+    >> AssertionError: Tile(0, 0, 25) is not valid
+
+    assert tms.is_valid(0, 0, 25, strict=False)
+    >> UserWarning: TileMatrix not found for level: 25 - Creating values from TMS Scale.
+    ```
 
 ## 6.2.0 (2024-12-19)